refactor: modularize scene labels interaction
