+++ js/scene.js
@@ -1,0 +1,240 @@
+import * as THREE from 'https://unpkg.com/three@0.159.0/build/three.module.js';
+import { updateLabels } from './labels.js';
+
+export let scene;
+export let camera;
+export let renderer;
+export let meshes;
+
+export function initScene(container, fpsCounter) {
+  function setContainerSize() {
+    const aspect = 16 / 9;
+    let width = window.innerWidth;
+    let height = window.innerHeight;
+    if (width / height > aspect) {
+      width = height * aspect;
+    } else {
+      height = width / aspect;
+    }
+    container.style.width = `${width}px`;
+    container.style.height = `${height}px`;
+    return { width, height };
+  }
+
+  scene = new THREE.Scene();
+  const { width: initW, height: initH } = setContainerSize();
+  camera = new THREE.PerspectiveCamera(75, initW / initH, 0.1, 1000);
+  renderer = new THREE.WebGLRenderer({ antialias: true });
+  renderer.setSize(initW, initH);
+  renderer.setClearColor(0x000033);
+  renderer.shadowMap.enabled = true;
+  container.appendChild(renderer.domElement);
+
+  // Ground with custom height pattern
+  console.info('Generating ground geometry');
+  const groundSize = 32;
+  const groundGeo = new THREE.PlaneGeometry(groundSize, groundSize, groundSize, groundSize);
+  const pos = groundGeo.attributes.position;
+  for (let i = 0; i < pos.count; i++) {
+    const height = i % 2 === 0 ? 0 : -1;
+    pos.setZ(i, height);
+  }
+  groundGeo.computeVertexNormals();
+
+  const groundMat = new THREE.MeshStandardMaterial({ color: 0x222222 });
+  const ground = new THREE.Mesh(groundGeo, groundMat);
+  ground.rotation.x = -Math.PI / 2;
+  ground.position.y = 0;
+  ground.receiveShadow = true;
+  scene.add(ground);
+  console.info('Ground added', ground.position);
+
+  // Lighting
+  const ambientLight = new THREE.AmbientLight(0xffffff, 0.3);
+  scene.add(ambientLight);
+  console.info('Ambient light added');
+
+  const dirLight = new THREE.DirectionalLight(0xffffff, 0.9);
+  dirLight.position.set(5, 10, 5);
+  dirLight.castShadow = true;
+  dirLight.shadow.mapSize.width = 2048;
+  dirLight.shadow.mapSize.height = 2048;
+  dirLight.shadow.camera.left = -15;
+  dirLight.shadow.camera.right = 15;
+  dirLight.shadow.camera.top = 15;
+  dirLight.shadow.camera.bottom = -15;
+  dirLight.shadow.camera.near = 1;
+  dirLight.shadow.camera.far = 50;
+  scene.add(dirLight);
+  console.info('Directional light added');
+
+  meshes = [];
+  function createMesh(geometry, color, x, z = 1) {
+    const material = new THREE.MeshStandardMaterial({ color });
+    const mesh = new THREE.Mesh(geometry, material);
+    mesh.position.set(x, 2, z);
+    mesh.castShadow = true;
+    mesh.receiveShadow = true;
+    scene.add(mesh);
+    meshes.push(mesh);
+    return mesh;
+  }
+
+  createMesh(new THREE.IcosahedronGeometry(1.2), 0xff6600, -4);
+  createMesh(new THREE.TorusGeometry(0.9, 0.3, 16, 30), 0x0096d6, 0);
+  console.info('Meshes created', meshes.map(m => m.position));
+
+  // Chunky voxel-style DEMOS heading
+  const LETTERS = {
+    D: ['11110', '10001', '10001', '10001', '10001', '10001', '11110'],
+    E: ['11111', '10000', '10000', '11110', '10000', '10000', '11111'],
+    M: ['10001', '11011', '10101', '10001', '10001', '10001', '10001'],
+    O: ['01110', '10001', '10001', '10001', '10001', '10001', '01110'],
+    S: ['01111', '10000', '10000', '01110', '00001', '00001', '11110'],
+    N: ['10001', '11001', '10101', '10011', '10001', '10001', '10001'],
+    T: ['11111', '00100', '00100', '00100', '00100', '00100', '00100'],
+    W: ['10001', '10001', '10001', '10101', '10101', '10101', '01010'],
+    H: ['10001', '10001', '10001', '11111', '10001', '10001', '10001'],
+    R: ['11110', '10001', '10001', '11110', '10100', '10010', '10001']
+  };
+
+  const textCubes = [];
+  const textLetters = [];
+
+  function createVoxelText(text) {
+    const size = 0.4;
+    const depth = 0.4;
+    const colors = [0xff0000, 0xff5800, 0xffd500, 0x009b48, 0x0045ad, 0xffffff];
+    const availableColors = [...colors];
+    const group = new THREE.Group();
+    let offsetX = 0;
+    text.toUpperCase().split('').forEach(ch => {
+      const pattern = LETTERS[ch];
+      if (!pattern) {
+        offsetX += size * 6;
+        return;
+      }
+      const letterGroup = new THREE.Group();
+      if (availableColors.length === 0) {
+        availableColors.push(...colors);
+      }
+      const color = availableColors.splice(Math.floor(Math.random() * availableColors.length), 1)[0];
+      const letterMaterial = new THREE.MeshStandardMaterial({ color });
+      pattern.forEach((row, y) => {
+        row.split('').forEach((bit, x) => {
+          if (bit === '1') {
+            const cube = new THREE.Mesh(new THREE.BoxGeometry(size, size, depth), letterMaterial);
+            cube.position.set(x * size, (pattern.length - y - 1) * size, 0);
+            cube.castShadow = true;
+            cube.receiveShadow = true;
+            cube.userData.phase = Math.random() * Math.PI * 2;
+            cube.userData.rotSpeed = new THREE.Vector3(0, 0, 0);
+            letterGroup.add(cube);
+            textCubes.push(cube);
+          }
+        });
+      });
+      const letterBox = new THREE.Box3().setFromObject(letterGroup);
+      const letterCenter = letterBox.getCenter(new THREE.Vector3());
+      const letterWidth = letterBox.getSize(new THREE.Vector3()).x;
+      letterGroup.children.forEach(c => {
+        c.position.sub(letterCenter);
+        c.userData.initialX = c.position.x;
+        c.userData.initialZ = c.position.z;
+      });
+      letterGroup.position.x = offsetX + letterWidth / 2;
+      group.add(letterGroup);
+      textLetters.push(letterGroup);
+      offsetX += letterWidth + size;
+    });
+    const box = new THREE.Box3().setFromObject(group);
+    const center = box.getCenter(new THREE.Vector3());
+    group.children.forEach(c => c.position.sub(center));
+    textLetters.forEach(letter => {
+      letter.userData.initialZ = letter.position.z;
+      letter.userData.phase = Math.random() * Math.PI * 2;
+    });
+    group.scale.set(2 / 3, 2 / 3, 2 / 3);
+    group.position.set(0, 4.35, 0.5);
+    group.rotation.x = 0;
+    return group;
+  }
+
+  const textMesh = createVoxelText('DEMOS');
+  scene.add(textMesh);
+  console.info('Voxel text added', textMesh.position);
+
+  camera.position.set(0, 7, 5);
+  camera.lookAt(0, 2, 0);
+
+  let lastTime;
+  let frames = 0;
+
+  function animate(timestamp) {
+    requestAnimationFrame(animate);
+    if (lastTime === undefined) lastTime = timestamp;
+    frames++;
+    if (timestamp - lastTime >= 1000) {
+      const fps = Math.round((frames * 1000) / (timestamp - lastTime));
+      if (fpsCounter) fpsCounter.textContent = fps + ' FPS';
+      frames = 0;
+      lastTime = timestamp;
+    }
+    meshes.forEach(mesh => {
+      mesh.rotation.x += 0.005;
+      mesh.rotation.y += 0.01;
+    });
+    textCubes.forEach(cube => {
+      const { rotSpeed, initialX, phase } = cube.userData;
+      cube.rotation.x += rotSpeed.x;
+      cube.rotation.y += rotSpeed.y;
+      cube.rotation.z += rotSpeed.z;
+      cube.position.x = initialX + Math.sin(timestamp / 600 + phase) * 0.05;
+    });
+    textLetters.forEach(letter => {
+      const { initialZ, phase } = letter.userData;
+      letter.position.z = initialZ + Math.sin(timestamp / 600 + phase) * 0.05;
+    });
+    updateLabels(camera, timestamp);
+    renderer.render(scene, camera);
+  }
+
+  function onWindowResize() {
+    const { width, height } = setContainerSize();
+    camera.aspect = width / height;
+    camera.updateProjectionMatrix();
+    renderer.setSize(width, height);
+    checkOrientation();
+    console.info('Window resized', { width, height });
+  }
+
+  function checkOrientation() {
+    const isLandscape = window.matchMedia('(orientation: landscape)').matches;
+    const isMobile = window.innerWidth <= 768;
+    if (isMobile && isLandscape) {
+      if (!document.fullscreenElement && container.requestFullscreen) {
+        container.requestFullscreen();
+      }
+    } else {
+      if (document.fullscreenElement && document.exitFullscreen) {
+        document.exitFullscreen();
+      }
+    }
+    console.info('Orientation checked', { isLandscape, isMobile });
+  }
+
+  window.addEventListener('resize', onWindowResize);
+  window.addEventListener('orientationchange', checkOrientation);
+  window.addEventListener('pointerup', checkOrientation);
+
+  function updateViewportHeight() {
+    document.documentElement.style.setProperty('--vh', `${window.innerHeight * 0.01}px`);
+  }
+  window.addEventListener('resize', updateViewportHeight);
+  window.addEventListener('orientationchange', updateViewportHeight);
+  updateViewportHeight();
+  onWindowResize();
+  checkOrientation();
+
+  requestAnimationFrame(animate);
+}
